requirements.txt
fastapi[all]
aiohttp
python-dotenv
pipecat-ai[daily]
